refactor: unify bounded Skill hit rules (#973)
